LPS-198392 Remove service from FieldProvider, rename it to FieldProviderUtil and pass objectDefinitionLocalService, objectEntryOpenAPIResourceProvider and openAPIYAMLProvider as a parameter
