+++ modules/apps/batch-planner/batch-planner-rest-impl/src/main/java/com/liferay/batch/planner/rest/internal/resource/v1_0/PlanResourceImpl.java
@@ -12,11 +12,14 @@ import com.liferay.batch.planner.model.BatchPlannerPolicy;
 import com.liferay.batch.planner.rest.dto.v1_0.Mapping;
 import com.liferay.batch.planner.rest.dto.v1_0.Plan;
 import com.liferay.batch.planner.rest.dto.v1_0.Policy;
-import com.liferay.batch.planner.rest.internal.vulcan.batch.engine.FieldProvider;
+import com.liferay.batch.planner.rest.internal.vulcan.batch.engine.util.FieldProviderUtil;
+import com.liferay.batch.planner.rest.internal.vulcan.yaml.openapi.OpenAPIYAMLProvider;
 import com.liferay.batch.planner.rest.resource.v1_0.PlanResource;
 import com.liferay.batch.planner.service.BatchPlannerMappingService;
 import com.liferay.batch.planner.service.BatchPlannerPlanService;
 import com.liferay.batch.planner.service.BatchPlannerPolicyService;
+import com.liferay.object.rest.openapi.v1_0.ObjectEntryOpenAPIResourceProvider;
+import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -71,8 +74,10 @@ public class PlanResourceImpl extends BasePlanResourceImpl {
 
 		return _getResponse(
 			TaskItemUtil.getSimpleClassName(internalClassNameKey),
-			_fieldProvider.getFields(
+			FieldProviderUtil.getFields(
 				contextCompany.getCompanyId(), internalClassNameKey,
+				_objectDefinitionLocalService,
+				_objectEntryOpenAPIResourceProvider, _openAPIYAMLProvider,
 				contextUriInfo));
 	}
 
@@ -140,7 +145,7 @@ public class PlanResourceImpl extends BasePlanResourceImpl {
 	}
 
 	private Response _getResponse(String dtoEntityName, List<Field> fields) {
-		fields = _fieldProvider.filter(fields, Field.AccessType.READ);
+		fields = FieldProviderUtil.filter(fields, Field.AccessType.READ);
 
 		Iterator<Field> iterator = fields.iterator();
 
@@ -244,6 +249,13 @@ public class PlanResourceImpl extends BasePlanResourceImpl {
 	private BatchPlannerPolicyService _batchPlannerPolicyService;
 
 	@Reference
-	private FieldProvider _fieldProvider;
+	private ObjectDefinitionLocalService _objectDefinitionLocalService;
+
+	@Reference
+	private ObjectEntryOpenAPIResourceProvider
+		_objectEntryOpenAPIResourceProvider;
+
+	@Reference
+	private OpenAPIYAMLProvider _openAPIYAMLProvider;
 
 }
